requirements_test.txt
nose==1.3.7
nose-cprof==0.2.1
nose-timer==0.7.1

# For plotting
matplotlib==2.1.2
moviepy==0.2.3.2
Pillow==5.0.0

# For Keras example
Keras==2.1.4
scikit-learn==0.18.0

# For movielens dataset
lightfm==1.13
